training_utils: split Kimi VL placeholder expansion into mm_data.py
data.py was growing with VL-only logic; move the 5 helpers and
expand_multimodal_rollout_data_in_place into mm_data.py so generic
batching stays out of multimodal preprocessing's way and future VL
models can extend mm_data.py without inflating data.py.

Pure mechanical refactor: function bodies are unchanged. Hardcoded
media token id 163605 and 2x2 patch merger dims are promoted to
KIMI_VL_MEDIA_TOKEN_ID / _KIMI_VL_MERGE_H / _KIMI_VL_MERGE_W constants.
data.py keeps a re-export of expand_multimodal_rollout_data_in_place
via 'from .mm_data import ...', so the two call sites in get_batch /
get_data_iterator are unchanged.
